% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firststep.R
\name{impute_mar}
\alias{impute_mar}
\title{Impute missing data under Missing At Random using mice or jomo}
\usage{
impute_mar(data, mi = c("mice", "jomo"), m = 5, seed = 123, ...)
}
\arguments{
\item{data}{A data frame containing missing values to be imputed.}

\item{mi}{A character string specifying the imputation method: "mice" or "jomo".}

\item{m}{Number of imputations to perform (default is 5).}

\item{seed}{A seed for reproducibility (default is 123).}

\item{...}{Additional arguments passed to the imputation functions.}
}
\value{
A list containing imputed datasets (\code{mice} output) or a single completed dataset (\code{jomo} output).
}
\description{
This function imputes missing data in a dataset using either the \code{mice} or \code{jomo} package.
}
\examples{

data("simdaNA")

summary(simdaNA)

library(mice)

# Imputation with mice

imputed_mice <- impute_mar(simdaNA[, c("Y","X1.mnar","X2")], mi = "mice",
method = c("logreg", "polr", "polyreg"), m = 10,printFlag = FALSE)

head(mice::complete(imputed_mice,1)) # imputation with mice



# Imputation with jomo

data("simda2NA")

summary(simda2NA)

library(jomo)


nimp <- 5
nburn <- 10
nbetween <- 10
formula = y ~ x1.mnar + x2 + (1|clus)


imputed_jomo <-  jomo.glmer( formula = as.formula(formula),
data = simda2NA[, c("y", "x1.mnar", "x2", "clus")], nburn = nburn,
nbetween = nbetween, nimp = nimp, output = 0)

imputed_jomo[2050:2056,]

}
