% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simda2NA.R
\docType{data}
\name{simda2NA}
\alias{simda2NA}
\title{Hierarchical data under missing not at random.}
\format{
A data.frame with 2000 rows and 6 columns:
\describe{
\item{id}{Unique identifier for each observation.}
\item{y}{Binary response variable (factor).}
\item{x1}{Ordered categorical variable.}
\item{x2}{Categorical variable (factor).}
\item{clus}{Cluster identifier.}
\item{x1.mis}{Ordered categorical variable with missing values.}
}
}
\source{
The data was generated using the \code{simmnar2} function on the \code{simda2} data.
}
\usage{
simda2NA
}
\description{
The data \code{simda2NA} contains  missing not at random (MNAR) values in the
hierarchical data.
}
\keyword{datasets}
