% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmnar.R
\name{simmnar}
\alias{simmnar}
\title{Simulation of MNAR Mechanism on ordinal variable}
\usage{
simmnar(data, Y, id = NULL, ord_var, A, Prob.A, B, Prob.B)
}
\arguments{
\item{data}{A data.frame.}

\item{Y}{The name of the binary variable (0 or 1) used as a conditioning criterion.}

\item{id}{An optional vector of identifiers. If NULL, a sequential index is used.}

\item{ord_var}{The name of the ordinal variable where missing values will be introduced.}

\item{A}{The levels of \code{ord_var} affected by the Group A mechanism.}

\item{Prob.A}{The probability of introducing missing values for observations in Group \code{A}.}

\item{B}{The levels of \code{ord_var} affected by the Group B mechanism.}

\item{Prob.B}{The probability of introducing missing values for observations in Group \code{B}.}
}
\value{
The \code{data.frame} with a new column containing the modified \code{ord_var},
now including missing values.
}
\description{
This function simulates a Missing Not At Random (MNAR) mechanism
by introducing missing values into an ordinal variable, based on
a binary response variable.
}
\examples{
# Example 1

set.seed(123) # Ensure reproducibility
simu <- data.frame(id = 1:1000,  Y = rbinom(1000, 1, 0.5),
          X1 = ordered(sample(1:5, 1000, replace = TRUE)),
          X2 = sample(letters[1:4], 1000, replace = TRUE))

head(simu)

simuNA <- simmnar( data = simu, Y = "Y",  id = "id", ord_var = "X1",
                         A = 1,  Prob.A = 0.3, B = 5,  Prob.B = 0.5)

head(simuNA)

# Example 2

data("simda")
head(simda)
simdaNA <- simmnar( data = simda, Y = "Y",  id = "id",
ord_var = "X1", A = 2,  Prob.A = 0.5, B = 4,  Prob.B = 0.8)

head(simdaNA)

summary(simdaNA)
}
\keyword{MNAR,}
\keyword{data}
\keyword{missing}
\keyword{ordinal}
\keyword{variable,}
