% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bin_nonhier.R
\name{simulate_bin_nonhier}
\alias{simulate_bin_nonhier}
\title{Simulate a Non-Hierarchical Dataset with a Binary Response}
\usage{
simulate_bin_nonhier(levels_X2 = 4, levels_X1 = 5, n = 1000, seed = NULL)
}
\arguments{
\item{levels_X2}{Integer. Number of levels for the X2 factor (default is 4).
Must be between 2 and 26.}

\item{levels_X1}{Integer. Number of ordered levels for the X1 factor (default is 5).}

\item{n}{Integer. Number of observations to simulate (default is 1000).}

\item{seed}{Integer or NULL. Random seed for reproducibility (default is NULL).}
}
\value{
A \code{data.frame} with the following columns:
\describe{
\item{id}{Integer identifier for each observation}
\item{X2}{Factor variable with alphabetic levels (e.g., a, b, c, ...)}
\item{X1}{Ordered factor with levels 1 < 2 < ...}
\item{Y}{Binary factor representing the outcome (0 or 1)}
}
}
\description{
Generates a simulated dataset with an ordinal predictor and a binary outcome.
The covariate X2 is a factor based on alphabetic levels (e.g., a, b, c, ...),
and X1 is an ordered factor with user-specified levels.
}
\examples{
# Simulate 1000 observations with 4 levels of X2 and 5 levels of X1
dat1 <- simulate_bin_nonhier(n = 1000, levels_X2 = 4, levels_X1 = 5, seed = 123)
head(dat1)

# Data struture
str(dat1)
}
