% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thirdstep_mice.R
\name{thirdstep_mice}
\alias{thirdstep_mice}
\title{Perform Rubin's rule}
\usage{
thirdstep_mice(data, formula = NULL, manydelta = NULL)
}
\arguments{
\item{data}{A dataset in the format of a \code{mitml.list}, containing the imputed
datasets.}

\item{formula}{A character string representing the model formula for MAR analysis.
Required if performing MAR analysis.}

\item{manydelta}{A data frame or matrix where each column represents a
variable to be used in MNAR analysis. Required if performing MNAR analysis.}
}
\value{
If \code{formula} is provided, returns the pooled model summary for the MAR analysis.
If \code{manydelta} is provided, returns a list of pooled model summaries for the MNAR analysis.
}
\description{
This function  performs either MAR (Missing At Random) or
MNAR (Missing Not At Random) analysis based on the provided inputs.
It fits generalized linear models (GLMs) with a logistic link function and
pools the results using the Rubin's rule.
}
\details{
\itemize{
\item For MAR analysis, a single GLM is fitted using the specified formula.
\item For MNAR analysis, a GLM is fitted for each variable in \code{manydelta}.
}
}
