% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firststep.R
\name{firststep}
\alias{firststep}
\title{Impute Missing Data under Missing At Random using mice or jomo}
\usage{
firststep(data, mi = c("mice", "jomo"), ...)
}
\arguments{
\item{data}{A dataframe containing missing values to be imputed.}

\item{mi}{A character string specifying the imputation method: either \code{"mice"} or \code{"jomo"}.}

\item{...}{Additional arguments passed to the respective imputation functions.}
}
\value{
For \code{mi = "mice"}, returns a \code{mids} object containing multiple imputations.
For \code{mi = "jomo"}, returns a single completed dataset or a list, depending on the
arguments and output type of \code{jomo}.
}
\description{
This function performs imputation of missing data in a dataset using either
the \code{mice} or \code{jomo} package, depending on the specified method. The \code{mice}
package uses multiple imputation by chained equations, while \code{jomo} employs
a multilevel imputation approach.
}
\examples{



data("simdaNA")
summary(simdaNA)
library(mice)
simdaNA$X1.mar <- simdaNA$X1.mis
imputed_mice <- firststep(
  data = simdaNA[, c("Y", "X1.mar", "X2")],
  mi = "mice",
  method = c("logreg", "polr", "polyreg"),
  m = 10,
  printFlag = FALSE
)
head(complete(imputed_mice, 1)) # View the first imputed dataset




# Example 2: Imputation with jomo
data("simda2NA")
summary(simda2NA)
library(jomo)

simda2NA$x1.mar <- simda2NA$x1.mis
formula <- y ~ x1.mar + x2 + (1 | clus)
imputed_jomo <- firststep(
  data = simda2NA[, c("y", "x1.mar", "x2", "clus")],
  mi = "jomo",
  formula = formula,
  nimp = 5,
  nburn = 10,
  nbetween = 10,
  output=0
)

head((mitml::jomo2mitml.list(imputed_jomo))[[1]]) # Convert jomo output to mitml format

}
