% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firststep.R
\name{impute_mar}
\alias{impute_mar}
\title{Impute missing data under Missing At Random using mice or jomo}
\usage{
impute_mar(data, mi = c("mice", "jomo"), m = 5, seed = 123, ...)
}
\arguments{
\item{data}{A data frame containing missing values to be imputed.}

\item{mi}{A character string specifying the imputation method: "mice" or "jomo".}

\item{m}{Number of imputations to perform (default is 5).}

\item{seed}{A seed for reproducibility (default is 123).}

\item{...}{Additional arguments passed to the imputation functions.}
}
\value{
A list containing imputed datasets (\code{mice} output) or a single completed dataset (\code{jomo} output).
}
\description{
This function imputes missing data in a dataset using either the \code{mice} or \code{jomo} package.
}
\examples{

data("simda")

simdaNA <- simmnar(simda, Y = "Y", id = "id", ord_var = "X1",
                   A = 1, Prob.A = 0.5, B = 5, Prob.B = 0.8)

library(mice)

# Imputation with mice
imputed_mice <- impute_mar(simdaNA[, c("Y","X1.mis","X2")], mi = "mice",
method = c("logreg", "polr", "polyreg"), m = 10,printFlag = FALSE)

head(mice::complete(imputed_mice,1)) # imputation with mice

library(jomo)
# Imputation with jomo
imputed_jomo <- impute_mar(data = simdaNA[, c("Y","X1.mis","X2")], mi = "jomo",
m = 5, output=0)

imputed_jomo[1504:1510,] #  imputation with jomo

}
