% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simda.R
\docType{data}
\name{simda}
\alias{simda}
\title{Non hierarchical data}
\format{
A data frame with 1000 rows and 4 columns:
\describe{
\item{id}{Unique identifier for observations.}
\item{Y}{Binary response variable, with values 0 and 1.}
\item{X1}{Ordinal variable with 5 categories (1 to 5).}
\item{X2}{Categorical variable with 4 categories (1 to 4).}
}
}
\source{
Generated with the script located in \code{data-raw/simda.R}.
}
\usage{
simda
}
\description{
\code{simda} is an example of a non hierarchical data.
}
\keyword{datasets}
