% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simda2.R
\docType{data}
\name{simda2}
\alias{simda2}
\title{Hierarchical data}
\format{
A data.frame with 2000 rows and 5 columns:
\describe{
\item{id}{Unique identifier for each observation.}
\item{y}{Binary response variable (factor).}
\item{x1}{Ordered categorical variable.}
\item{x2}{Categorical variable (factor).}
\item{clus}{Cluster identifier.}
}
}
\source{
Generated with the script located in\code{data-raw/simda2.R}
}
\usage{
simda2
}
\description{
\code{simda2} is an example of a hierarchical data.
}
\keyword{datasets}
