% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmnar.R
\name{simmnar}
\alias{simmnar}
\title{Simulate a Missing Not At Random (MNAR) Mechanism in Non-Hierarchical Data}
\usage{
simmnar(data, Y, ord_var, A, probA, B, probB, verbose = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} containing the data.}

\item{Y}{The name of the binary outcome variable (must contain 0 and 1 only).}

\item{ord_var}{The name of the ordinal variable where missingness is introduced.}

\item{A}{Levels of \code{ord_var} for which missingness is introduced when \code{Y == 1}.}

\item{probA}{Probability of missingness in group A (between 0 and 1).}

\item{B}{Levels of \code{ord_var} for which missingness is introduced when \code{Y == 0}.}

\item{probB}{Probability of missingness in group B (between 0 and 1).}

\item{verbose}{Logical. If \code{TRUE}, prints the number of introduced missing values (default = TRUE).}
}
\value{
A \code{data.frame} with a new column \code{<ord_var>.mis} containing missing values.
}
\description{
This function introduces missing values into an ordinal variable based on
the response variable \code{Y} (binary), creating a MNAR mechanism with
specified probabilities.
}
\examples{
if (requireNamespace("ovsa", quietly = TRUE)) {
  data(simda)

  simdaNA <- simmnar(
    data = simda,
    Y = "Y",
    ord_var = "X1",
    A = 1,
    probA = 0.3,
    B = 5,
    probB = 0.5
  )

  summary(simdaNA)
}
}
