% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmnar.R
\name{simmnar}
\alias{simmnar}
\title{Simulate a Missing Not At Random (MNAR) mechanism in the Non-Hierarchical data}
\usage{
simmnar(data, Y, id = NULL, ord_var, A, probA, B, probB)
}
\arguments{
\item{data}{A data.frame.}

\item{Y}{The name of the binary variable (0 or 1) used as a conditioning criterion.}

\item{id}{An optional vector of identifiers. If NULL, a sequential index is used.}

\item{ord_var}{The name of the ordinal variable where missing values will be introduced.}

\item{A}{The levels of \code{ord_var} affected by the Group A mechanism.}

\item{probA}{The probability of introducing missing values for observations in Group \code{A}.}

\item{B}{The levels of \code{ord_var} affected by the Group B mechanism.}

\item{probB}{The probability of introducing missing values for observations in Group \code{B}.}
}
\value{
A \code{data.frame} with a new column containing the modified \code{ord_var},
now including missing values.
}
\description{
This function simulates a MNAR mechanism
by introducing missing values into an independent ordinal variable with
specific probabilities,  based on a binary response variable.
}
\examples{
#  Example 1

set.seed(123) # for reproducibility
simu <- data.frame(id = 1:1000,  Y = rbinom(1000, 1, 0.5),
          X1 = ordered(sample(1:5, 1000, replace = TRUE)),
          X2 = sample(letters[1:4], 1000, replace = TRUE))

head(simu)

simuNA <- simmnar( data = simu, Y = "Y",  id = "id", ord_var = "X1",
                         A = 1,  probA = 0.3, B = 5,  probB = 0.5)# use simmnar

head(simuNA)



#  Example 2

#' set.seed(321) # for reproducibility
data("simda")
head(simda)
simdaNA <- simmnar( data = simda, Y = "Y",  id = "id",
ord_var = "X1", A = 2,  probA = 0.5, B = 4,  probB = 0.8) # use simmnar

head(simdaNA)

summary(simdaNA)
}
\keyword{Context.}
\keyword{MNAR,}
\keyword{Non-Hierarchical}
\keyword{Simulation,}
\keyword{data,}
\keyword{missing}
