% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmnar2.R
\name{simmnar2}
\alias{simmnar2}
\title{Simulation a Missing Not At Random (MNAR) Mechanism in the Hierarchical data}
\usage{
simmnar2(data, proba, cat_var, Y, id, ord_var, A, B)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{proba}{A named list or data frame containing the probabilities of missingness
for each stratum of \code{cat_var}. Each stratum must contain a vector of two probabilities:
one for group A and one for group B.}

\item{cat_var}{The name of the categorical variable (stratification variable).}

\item{Y}{The name of the binary variable indicating group membership (1 for group A, 0 for group B).}

\item{id}{The name of the variable identifying unique individuals.}

\item{ord_var}{The name of the ordinal variable where missingness will be introduced.}

\item{A}{A vector of values in \code{ord_var} defining group A.}

\item{B}{A vector of values in \code{ord_var} defining group B.}
}
\value{
A data frame with a new column named \verb{<ord_var>.mis}, containing the ordinal variable
with missing values introduced under the MNAR mechanism.
}
\description{
This function simulates a Missing Not At Random (MNAR) mechanism by introducing
missing values into an independent ordinal variable with specific probabilities
based on a binary response variable and a independent categorical variable.
}
\examples{
#-------------------------  Example  ------------------------------------#

 data("simda2")
 head(simda2)

 missing_prob <- data.frame(matrix(c(0.2,0.3,0.5,0.4,0.4,0.2,0.3,0.3),
 nrow = 2, byrow = FALSE,dimnames = list(c("0","1"),paste0("proba",1:4))))
 missing_prob

 set.seed(215) # for reproducibility
 simda2NA <- simmnar2(data = simda2,proba = missing_prob,cat_var = "x2",
 Y = "y", id = "id", ord_var = "x1", A = 1, B = 3)

 simda2NA[102:108,]

 summary(simda2NA)

}
\keyword{Context.}
\keyword{Hierarchical}
\keyword{MNAR,}
\keyword{Simulation,}
\keyword{data,}
\keyword{missing}
